userInfoStore 분리, 경로 수정
