+++ client/src/components/ModalComponents/TaskModal/TaskModal.tsx
@@ -26,7 +26,8 @@ import dayjs from "dayjs";
 import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs";
 import { LocalizationProvider } from "@mui/x-date-pickers/LocalizationProvider";
 import { TimePicker } from "@mui/x-date-pickers/TimePicker";
-import { toggleTaskModal, useUserInfoStore } from "../../../store/store";
+import { toggleTaskModal } from "../../../store/store";
+import { useUserInfoStore } from "../../../store/userInfoStore";
 
 const TaskModal = () => {
   const selectedDate = useUserInfoStore((state) => state.selectedDate);
